+++ src/main/java/br/com/alelo/consumerpat/core/dto/v1/response/AddressV1ResponseDto.java
@@ -1,4 +1,4 @@
-package br.com.alelo.consumerpat.core.v1.response;
+package br.com.alelo.consumerpat.core.dto.v1.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
