+++ src/main/java/br/com/alelo/consumerpat/core/dto/v1/response/ContactV1ResponseDto.java
@@ -1,13 +1,15 @@
-package br.com.alelo.consumerpat.core.v1.request;
+package br.com.alelo.consumerpat.core.dto.v1.response;
 
-import lombok.Data;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Data
+@Builder
+@AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class ContactV1RequestDto {
+public class ContactV1ResponseDto {
 
     private String mobilePhone;
     private String residencePhone;
