+++ src/main/java/br/com/alelo/consumerpat/core/exception/BadRequestException.java
@@ -11,4 +11,8 @@ public class BadRequestException extends Exception {
     public BadRequestException(Map<String, String> fieldsAndMessages) {
         this.fieldsAndMessages = fieldsAndMessages;
     }
+
+    public Map<String, String> getFieldsAndMessages() {
+        return fieldsAndMessages;
+    }
 }
