+++ src/main/java/br/com/alelo/consumerpat/core/mapper/domain/ConsumerDomainMapper.java
@@ -1,8 +1,8 @@
 package br.com.alelo.consumerpat.core.mapper.domain;
 
 import br.com.alelo.consumerpat.core.domain.ConsumerDomain;
-import br.com.alelo.consumerpat.core.v1.request.ConsumerCreateV1RequestDto;
-import br.com.alelo.consumerpat.core.v1.request.ConsumerUpdateV1RequestDto;
+import br.com.alelo.consumerpat.core.dto.v1.request.ConsumerCreateV1RequestDto;
+import br.com.alelo.consumerpat.core.dto.v1.request.ConsumerUpdateV1RequestDto;
 
 public class ConsumerDomainMapper {
