+++ src/main/java/br/com/alelo/consumerpat/core/mapper/entity/ConsumerEntityMapper.java
@@ -2,7 +2,7 @@ package br.com.alelo.consumerpat.core.mapper.entity;
 
 import br.com.alelo.consumerpat.core.dataprovider.entity.ConsumerEntity;
 import br.com.alelo.consumerpat.core.domain.ConsumerDomain;
-import br.com.alelo.consumerpat.core.v1.request.ConsumerUpdateV1RequestDto;
+import br.com.alelo.consumerpat.core.dto.v1.request.ConsumerUpdateV1RequestDto;
 import br.com.alelo.consumerpat.core.mapper.domain.ConsumerDomainMapper;
 
 public class ConsumerEntityMapper {
