+++ src/main/java/br/com/alelo/consumerpat/core/mapper/response/ContactV1ResponseMapper.java
@@ -1,7 +1,7 @@
 package br.com.alelo.consumerpat.core.mapper.response;
 
 import br.com.alelo.consumerpat.core.dataprovider.entity.ContactEntity;
-import br.com.alelo.consumerpat.core.v1.response.ContactV1ResponseDto;
+import br.com.alelo.consumerpat.core.dto.v1.response.ContactV1ResponseDto;
 
 public class ContactV1ResponseMapper {
