+++ src/main/java/br/com/alelo/consumerpat/core/usecase/ConsumerFindUseCase.java
@@ -1,8 +1,8 @@
 package br.com.alelo.consumerpat.core.usecase;
 
-import br.com.alelo.consumerpat.core.v1.request.ConsumerPaginatedV1RequestDto;
-import br.com.alelo.consumerpat.core.v1.response.ConsumerV1ResponseDto;
-import br.com.alelo.consumerpat.core.v1.response.PaginatedResponseDto;
+import br.com.alelo.consumerpat.core.dto.v1.request.ConsumerPaginatedV1RequestDto;
+import br.com.alelo.consumerpat.core.dto.v1.response.ConsumerV1ResponseDto;
+import br.com.alelo.consumerpat.core.dto.v1.response.PaginatedResponseDto;
 
 public interface ConsumerFindUseCase {
