+++ src/main/java/br/com/alelo/consumerpat/core/usecase/ConsumerUpdateUseCase.java
@@ -1,6 +1,6 @@
 package br.com.alelo.consumerpat.core.usecase;
 
-import br.com.alelo.consumerpat.core.v1.request.ConsumerUpdateV1RequestDto;
+import br.com.alelo.consumerpat.core.dto.v1.request.ConsumerUpdateV1RequestDto;
 import br.com.alelo.consumerpat.core.exception.ConsumerNotFound;
 
 public interface ConsumerUpdateUseCase {
