+++ src/main/resources/messages.properties
@@ -1,6 +1,6 @@
 invalid.item=The {0} must not be null
-not_found.consumer=The consumer was not found
 invalid.balance=The balance is invalid
 invalid.already_registered=The document has already been registered
 invalid.recharge=The recharge must be greater than 0
 invalid.establishment_type_for_card=The Establishment has been informed is invalid for card
+error.server=The server encountered a problem
